chore style and layout

+++ db/migrate/20200328104207_create_ads.rb
@@ -3,7 +3,7 @@ class CreateAds < ActiveRecord::Migration[5.2]
     create_table :ads do |t|
       t.string :title, null: false
       t.text :description
-      t.numeric :total_price, null: false
+      t.integer :total_price, null: false
       t.references :user, foreign_key: true
 
       t.timestamps
