+++ db/migrate/20200406162025_create_kites.rb
@@ -4,14 +4,14 @@ class CreateKites < ActiveRecord::Migration[5.2]
       t.references :user, foreign_key: true
       t.string :brand, null: false
       t.string :name, null: false
-      t.numeric :year, null: false
-      t.numeric :size, null: false
+      t.integer :year, null: false
+      t.integer :size, null: false
       t.string :type
       t.string :sling_system
-      t.numeric :length_slim
+      t.integer :length_slim
       t.boolean :one_pump
-      t.numeric :price, null: false
-      t.numeric :quality, null: false
+      t.integer :price, null: false
+      t.integer :quality, null: false
       t.string :city
       t.boolean :cargoable
       t.string :origin_site
